Rebuild LiveCameras mechanism for auto restart, and fix threading
Cameras are now started and stopped from a single thread, which fixes
all the hidden race conditions that were previously lurking in here.

This also allows us to restart cameras when they go down (eg network
goes down for however long).

One side effect of all this is that the AddCamera/RemoveCamera/Reconfigure
mechanism is much simpler. You simply update the DB, and the LiveCameras
system takes care of it from there.