Include HD stream in camera liveness detection
Previously, we only monitored the LD stream for camera liveness.
But it just so happens, that 1 time out of 10, one of my cameras will
simply not send any HD packets. I don't know if this is a gortsplib
bug or a camera bug. Either way, we now detect this condition, and
restart the camera if either the LD or the HD stream stops sending
packets.